+++ burguer-queen-api-client/src/components/orders.tsx
@@ -1,7 +1,7 @@
-import { useState, useEffect } from 'react';
+//import { useState, useEffect } from 'react';
 import Header from './header';
 import MeseroNav from './meseroNavegacion';
-import ProductCard from './productCard'; // Importa el nuevo componente
+
 
 function Orders() {
   const containerStyle = {
@@ -14,52 +14,12 @@ function Orders() {
     padding: '20px',
   };
 
-  const [products, setProducts] = useState([]); // Estado para almacenar los productos
-  const token = localStorage.getItem('authToken');
-
-  useEffect(() => {
-    const fetchProducts = async () => {
-      try {
-        const response = await fetch('http://localhost:8080/products', {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: token,
-          },
-        });
-
-        if (response.ok) {
-          const data = await response.json();
-          console.log(data);
-          setProducts(data); // Guarda los productos en el estado
-        } else {
-          console.error('Error al cargar productos');
-        }
-      } catch (error) {
-        console.error('Error en la solicitud de productos:', error);
-      }
-    };
-
-    fetchProducts();
-  }, [token]);
 
   return (
     <>
       <Header />
       <MeseroNav />
       <div style={containerStyle}>
-        <div className="d-grid gap-4 "></div>
-        <div style={{ width: '40%' }}>
-          <input
-            placeholder="Cliente:"
-            style={{ backgroundColor: 'white', width: '100%', color: 'black' }}
-          />
-        </div>
-        <div style={{ display: 'flex', flexWrap: 'wrap', justifyContent: 'center' }}>
-          {products.map((product) => (
-            <ProductCard key={product.id} product={product} />
-          ))}
-        </div>
       </div>
     </>
   );
